requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-identity==1.15.0
azure-mgmt-core==1.6.0
azure-mgmt-storage==23.0.1
azure-mgmt-web==7.2.0
azure-storage-blob==12.19.0
blinker==1.9.0
boto3==1.34.0
botocore==1.34.162
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
ecdsa==0.19.1
fastapi==0.104.1
gitdb==4.0.12
GitPython==3.1.40
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
narwhals==2.1.2
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.27.0
streamlit==1.48.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
